+++ examples/full-jaffle-shop-demo/dbt/models/customers.sql
@@ -1,3 +1,5 @@
+{{ config(alias='customers_alias') }}
+
 with customers as (
 
     select * from {{ ref('stg_customers') }}