+++ packages/backend/src/database/entities/savedCharts.ts
@@ -67,6 +67,7 @@ export type DbSavedChart = {
     last_version_chart_kind: ChartKind;
     last_version_updated_at: Date;
     last_version_updated_by_user_uuid: string | undefined;
+    search_vector: string;
 };
 
 export type DbSavedChartVersion = {