+++ packages/frontend/src/features/scheduler/components/SchedulerForm.tsx
@@ -441,14 +441,8 @@ const SchedulerForm: FC<Props> = ({
                         {isThresholdAlert && (
                             <Stack spacing="xs">
                                 <FieldSelect
-                                    label={
-                                        <Text>
-                                            Alert field{' '}
-                                            <Text color="red" span>
-                                                *
-                                            </Text>{' '}
-                                        </Text>
-                                    }
+                                    label="Alert field"
+                                    required
                                     disabled={isThresholdAlertWithNoFields}
                                     withinPortal
                                     hasGrouping
@@ -488,7 +482,6 @@ const SchedulerForm: FC<Props> = ({
                                 )}
                                 <Group noWrap grow>
                                     <Select
-                                        required
                                         label="Condition"
                                         data={thresholdOperatorOptions}
                                         {...form.getInputProps(